Code review: own package
