+++ core/core-impl/src/main/java/ai/timefold/solver/core/api/score/stream/ConstraintCollectors.java
@@ -36,6 +36,7 @@ import ai.timefold.solver.core.api.function.ToLongTriFunction;
 import ai.timefold.solver.core.api.function.TriFunction;
 import ai.timefold.solver.core.api.function.TriPredicate;
 import ai.timefold.solver.core.api.score.stream.bi.BiConstraintCollector;
+import ai.timefold.solver.core.api.score.stream.common.SequenceChain;
 import ai.timefold.solver.core.api.score.stream.quad.QuadConstraintCollector;
 import ai.timefold.solver.core.api.score.stream.tri.TriConstraintCollector;
 import ai.timefold.solver.core.api.score.stream.uni.UniConstraintCollector;
@@ -1904,173 +1905,6 @@ public final class ConstraintCollectors {
     public static <A> UniConstraintCollector<A, ?, SequenceChain<A, Integer>>
             toConsecutiveSequences(ToIntFunction<A> indexMap) {
         return InnerUniConstraintCollectors.toConsecutiveSequences(indexMap);
-    }
-
-    /**
-     * Contains info regarding the consecutive sequences and breaks in a collection of points.
-     *
-     * @param <Value_> The type of value in the sequence.
-     * @param <Difference_> The type of difference between values in the sequence.
-     */
-    public interface SequenceChain<Value_, Difference_ extends Comparable<Difference_>> {
-
-        /**
-         * @return never null; the sequences contained in the collection in ascending order.
-         */
-        Iterable<Sequence<Value_, Difference_>> getConsecutiveSequences();
-
-        /**
-         * @return never null; the breaks contained in the collection in ascending order.
-         */
-        Iterable<Break<Value_, Difference_>> getBreaks();
-
-        /**
-         * Returns the first sequence of consecutive values.
-         *
-         * @return null if there are no sequences
-         */
-        Sequence<Value_, Difference_> getFirstSequence();
-
-        /**
-         * Returns the last sequence of consecutive values.
-         *
-         * @return null if there are no sequences
-         */
-        Sequence<Value_, Difference_> getLastSequence();
-
-        /**
-         * Returns the first break between two consecutive sequences of values.
-         *
-         * @return null if there are less than two sequences
-         */
-        Break<Value_, Difference_> getFirstBreak();
-
-        /**
-         * Returns the last break between two consecutive sequences of values.
-         *
-         * @return null if there are less than two sequences
-         */
-        Break<Value_, Difference_> getLastBreak();
-
-    }
-
-    /**
-     * Represents a series of consecutive values.
-     * For instance, the list [1,2,4,5,6,10] has three sequences: [1,2], [4,5,6], and [10].
-     *
-     * @param <Value_> The type of value in the sequence.
-     * @param <Difference_> The type of difference between values in the sequence.
-     */
-    public interface Sequence<Value_, Difference_ extends Comparable<Difference_>> {
-
-        /**
-         * @return never null; the first item in the sequence.
-         */
-        Value_ getFirstItem();
-
-        /**
-         * @return never null; the last item in the sequence.
-         */
-        Value_ getLastItem();
-
-        /**
-         * @return true if and only if this is the first sequence
-         */
-        boolean isFirst();
-
-        /**
-         * @return true if and only if this is the last sequence
-         */
-        boolean isLast();
-
-        /**
-         * @return If this is not the first sequence, the break before it. Otherwise, null.
-         */
-        Break<Value_, Difference_> getPreviousBreak();
-
-        /**
-         * @return If this is not the last sequence, the break after it. Otherwise, null.
-         */
-        Break<Value_, Difference_> getNextBreak();
-
-        /**
-         * @return never null; items in this sequence
-         */
-        Iterable<Value_> getItems();
-
-        /**
-         * @return the number of items in this sequence
-         */
-        int getCount();
-
-        /**
-         * @return never null; the difference between the last item and first item in this sequence
-         */
-        Difference_ getLength();
-    }
-
-    /**
-     * Represents a gap between two {@link Sequence sequences}.
-     * For instance, the list [1,2,4,5,6,10] has a break of length 2 between 2 and 4,
-     * as well as a break of length 4 between 6 and 10.
-     *
-     * @param <Value_> The type of value in the sequence.
-     * @param <Difference_> The type of difference between values in the sequence.
-     */
-    public interface Break<Value_, Difference_ extends Comparable<Difference_>> {
-
-        /**
-         * @return never null; the sequence leading directly into this
-         */
-        Sequence<Value_, Difference_> getPreviousSequence();
-
-        /**
-         * @return never null; the sequence immediately following this
-         */
-        Sequence<Value_, Difference_> getNextSequence();
-
-        /**
-         * @return true if and only if this is the first break
-         */
-        default boolean isFirst() {
-            return getPreviousSequence().isFirst();
-        }
-
-        /**
-         * @return true if and only if this is the last break
-         */
-        default boolean isLast() {
-            return getNextSequence().isLast();
-        }
-
-        /**
-         * Return the end of the sequence before this break. For the
-         * break between 6 and 10, this will return 6.
-         *
-         * @return never null; the item this break is directly after
-         */
-        default Value_ getPreviousSequenceEnd() {
-            return getPreviousSequence().getLastItem();
-        };
-
-        /**
-         * Return the start of the sequence after this break. For the
-         * break between 6 and 10, this will return 10.
-         *
-         * @return never null; the item this break is directly before
-         */
-        default Value_ getNextSequenceStart() {
-            return getNextSequence().getFirstItem();
-        }
-
-        /**
-         * Return the length of the break, which is the difference
-         * between {@link #getNextSequenceStart()} and {@link #getPreviousSequenceEnd()}. For the
-         * break between 6 and 10, this will return 4.
-         *
-         * @return never null; the length of this break
-         */
-        Difference_ getLength();
     }
 
     /**
