+++ core/core-impl/src/main/java/ai/timefold/solver/core/impl/score/stream/BreakImpl.java
@@ -2,11 +2,11 @@ package ai.timefold.solver.core.impl.score.stream;
 
 import java.util.function.BiFunction;
 
-import ai.timefold.solver.core.api.score.stream.ConstraintCollectors;
-import ai.timefold.solver.core.api.score.stream.ConstraintCollectors.Sequence;
+import ai.timefold.solver.core.api.score.stream.common.Break;
+import ai.timefold.solver.core.api.score.stream.common.Sequence;
 
 final class BreakImpl<Value_, Point_ extends Comparable<Point_>, Difference_ extends Comparable<Difference_>>
-        implements ConstraintCollectors.Break<Value_, Difference_> {
+        implements Break<Value_, Difference_> {
 
     private final BiFunction<Point_, Point_, Difference_> lengthFunction;
     private final SequenceImpl<Value_, Point_, Difference_> nextSequence;
