+++ core/core-impl/src/main/java/ai/timefold/solver/core/impl/score/stream/bi/InnerBiConstraintCollectors.java
@@ -21,8 +21,8 @@ import java.util.function.ToLongBiFunction;
 
 import ai.timefold.solver.core.api.function.QuadFunction;
 import ai.timefold.solver.core.api.function.TriFunction;
-import ai.timefold.solver.core.api.score.stream.ConstraintCollectors.SequenceChain;
 import ai.timefold.solver.core.api.score.stream.bi.BiConstraintCollector;
+import ai.timefold.solver.core.api.score.stream.common.SequenceChain;
 import ai.timefold.solver.core.impl.score.stream.ReferenceAverageCalculator;
 
 public class InnerBiConstraintCollectors {
