+++ core/core-impl/src/main/java/ai/timefold/solver/core/impl/score/stream/uni/ConsecutiveSequencesUniConstraintCollector.java
@@ -4,7 +4,7 @@ import java.util.Objects;
 import java.util.function.Supplier;
 import java.util.function.ToIntFunction;
 
-import ai.timefold.solver.core.api.score.stream.ConstraintCollectors.SequenceChain;
+import ai.timefold.solver.core.api.score.stream.common.SequenceChain;
 import ai.timefold.solver.core.impl.score.stream.SequenceCalculator;
 import ai.timefold.solver.core.impl.util.ConstantLambdaUtils;
